fix: fill doc_id as int64 so a batch can pass 256 documents
`BrendaDataset.__getitems__` filled each document's `doc_id` with its batch
position as a uint8 tensor, so the 257th document of a batch raised an
overflow RuntimeError. `TokenBudgetBatchSampler` caps a batch's padded
chunks, not its documents, so a large budget over short documents reaches
it. The fill is now int64 with the same values; downstream reads only the
tensor's length, the chunk count, so no model input changes.

+++ src/d3text/data/data.py
@@ -486,9 +486,11 @@ class BrendaDataset(Dataset):
             {
                 "id": self.data.iloc[ix]["pubmed_id"],
                 "sequence": seqdict[ix],
+                # Not uint8: `TokenBudgetBatchSampler` caps a batch's chunks,
+                # not its documents, so a position can pass 255.
                 "doc_id": torch.tensor(
                     [doc_id] * seqdict[ix]["input_ids"].shape[0],
-                    dtype=torch.uint8,
+                    dtype=torch.int64,
                 ),
                 "entities": self.data.iloc[ix]["entities"],
                 "relations": self.data.iloc[ix]["relations"],
